Styling and Semantic changes. Static build done
static build of site is basically all done. Changed fight avatars to
show who was in which corner. Added more buttons and color coded them
so you can pick which fighter in which corner will win and how.

Also on the index page changed the call to action kind of area to help
explain what the site is and how to use it.

Also changed all buttons to look much better for now.