got details of results working.
Got the percents to show properly using a weird way thats probably not
the correct or easiest way. I created ng-repeat divs that repeated thru
the results of a corner’s fight picks. it then filters then other non
results into a filtered array. called the remains of the filtered
array.length to find the rest of the votes for that fighter. Made a
specific, messy, controller for that series of divs. this will work for
now