Fixed chat room login catch. added comments
Now will display you as ‘guest’ if you are not logged in and you add to
the chatroom. Also, removes delete buttons if you are not logged in. i
also added a shit ton of comments to the html and js to explain what
the hell i was doing to my future self